fixing double promotion

+++ projects/fw_102_103/src/main.c
@@ -66,7 +66,7 @@ TASK(ads1115_writer, TASK_STACK_256) {
   while (true) {
     ads1115_read_converted(&ads1115_cfg, ADS1115_CHANNEL_0, &voltage);
     StatusCode status = queue_send(&ads1115_data_queue, &voltage, 1000);
-    LOG_DEBUG("Writing to ADC queue: %f\n", voltage);
+    LOG_DEBUG("Writing to ADC queue: %f\n", (double)voltage);
     if (status != STATUS_CODE_OK) {
       LOG_DEBUG("Write Failed\n");
     }
@@ -82,7 +82,7 @@ TASK(ads1115_reader, TASK_STACK_256) {
   while (true) {
     StatusCode status = queue_receive(&ads1115_data_queue, &voltage, 1000);
     if (status == STATUS_CODE_OK) {
-      LOG_DEBUG("Reading from ADC queue: %f\n", voltage);
+      LOG_DEBUG("Reading from ADC queue: %f\n",(double)voltage);
     } else {
       LOG_DEBUG("Read Failed\n");
     }
